+++ ecs/ecs.h
@@ -85,10 +85,9 @@ namespace Fresa::Component
     template <typename> struct component_tag {};
     
     template<typename C, std::enable_if_t<Reflection::is_in_variant<C, ComponentType>::value, bool> = true>
-    ComponentID getID() {
-        static ComponentID id_ = getVariantIndex<C, ComponentType>::value;
-        if (id_ >= MAX_COMPONENTS)
-            throw std::runtime_error("Increase max component capacity");
+    constexpr ComponentID getID() {
+        constexpr ComponentID id_ = getVariantIndex<C, ComponentType>::value;
+        static_assert(id_ < MAX_COMPONENTS, "Increase max component capacity");
         return id_;
     }